updatess....

+++ cumulants/bash/run_cumulants_sbi.sh
@@ -27,16 +27,33 @@ RUN_FROZEN=false
 USE_PLANCK=false
 
 DEFAULT_NDE_TYPE="CNF"
-DEFAULT_N_NDES=3
+DEFAULT_N_NDES=1 
+USE_SCALERS=true
 
+FORCE_FLAT_PRIOR=false
+FORCE_QUIJOTE_PRIOR=true
+NON_GAUSSIAN_TEST=false
+PLOT_FISHER_CLIPPED=true
+FORCE_RECOMPUTE_DATASET=false
+
+COMPRESSION="nn"
+USE_PRECISION_NN=false
+DATA_PROCESS_TYPE_NN="d"
+
+USE_SOBOL=true
 FORCE_NOISELESS_DATAVECTOR=false
 USE_QUIJOTE_TAILS=false # Use tails datavectors measured, not calculated, from Quijote
 FIDUCIAL_REDUCE=true # Reduce cumulants with fiducial variances
-USE_SCALERS=true # Not implemented yet
-
 DEFAULT_RESOLUTION=1024
 
-COMPRESSION="nn"
+RECALCULATE_DATASETS=false
+DATASETS_DIR="/project/ls-gruen/users/jed.homer/sbiaxpdf/quijote_data/datasets/"
+
+N_DATAVECTORS=10 # Number of independent datavectors to sample posteriors with (with N_SEEDS different posteriors)
+
+# Posterior sampling
+AFFINE_SAMPLE=false
+BLACKJAX_SAMPLE=true
 
 # Running a test single run or not
 if [[ "$SINGLE_RUN" == "true" ]]; then
@@ -46,40 +63,43 @@ if [[ "$SINGLE_RUN" == "true" ]]; then
     N_SEEDS_GLOBAL=1    # Number of repeated trainings for SBI 
     END_SEED=$(( $START_SEED + $N_SEEDS - 1 ))
     N_PARALLEL=2
-
-    RUN_FROZEN=false
 else
     echo "MULTIPLE SEEDS RUN."
-    N_SEEDS=50
+    N_SEEDS=10 # 100 # Number of independent datavectors to test each SBI with
     START_SEED=0
-    N_SEEDS_GLOBAL=10   # Number of repeated trainings for SBI
+    N_SEEDS_GLOBAL=1 # 9 # Number of repeated trainings for SBI
     END_SEED=$(( $START_SEED + $N_SEEDS - 1 ))
     N_PARALLEL=100
 fi
 
-N_GB=8
+N_GB=12
 N_CPU=8
-JOB_TIME="02:00:00"
+JOB_TIME="16:00:00"
 MAIL_TYPE="begin,end,fail"
 JOB_ARRAY_STR="$START_SEED-$END_SEED%$N_PARALLEL"
-
-N_DATAVECTORS=10       # Number of independent datavectors to sample posteriors with    
-N_LINEAR_SIMS=2000      # Number of linear simulations to use for training / pre-training
 
 order_idxs=(
     # "0"
     "0 1 2"
 )
 
-scales_sets=(
-    # "15.0 20.0 25.0 30.0 35.0"
-    "5.0 10.0 15.0 20.0 25.0 30.0 35.0"
-)
+if [ "$USE_SOBOL" == "true" ]; then
+    # SCALES="5.8 9.7 13.6 17.5 21.4 25.3 29.2 33.2"
+    scales_sets=(
+        "5.9 9.8 13.7 17.6 21.5 25.4 29.3 33.2"
+    )
+    N_LINEAR_SIMS=32768 
+else
+    scales_sets=(
+        "5.0 10.0 15.0 20.0 25.0 30.0 35.0" 
+    )
+    N_LINEAR_SIMS=2000 
+fi
 
 all_redshifts=(
     0.0 
-    # 0.5 
-    # 1.0
+    0.5 
+    1.0
 )
 
 # SBATCH out directory
@@ -88,9 +108,12 @@ OUT_DIR="/project/ls-gruen/users/jed.homer/sbiaxpdf/sbatch_outs/cumulants_sbi/$T
 mkdir -p "$OUT_DIR"
 
 # Empty datasets dir, recalculate them only once each
-DATASETS_DIR="/project/ls-gruen/users/jed.homer/sbiaxpdf/quijote_data/datasets/"
-rm -rf "${DATASETS_DIR:?}/"*
-echo "Emptied datasets directory: $DATASETS_DIR"
+if [ "$RECALCULATE_DATASETS" == true ]; then
+    rm -rf "${DATASETS_DIR:?}/"*                      # NOTE: not recalculating datasets here
+    echo "Emptied datasets directory: $DATASETS_DIR"
+else
+    echo "Didn't empty datasets directory: $DATASETS_DIR"
+fi
 
 # Base logging directory
 BASE_LOG_DIR="/project/ls-gruen/users/jed.homer/sbiaxpdf/$RESULTS_DIR/logs/"
@@ -107,8 +130,10 @@ fi
 
 # Run cumulants-from-quijote if using high resolution
 data_deps=()
+data_dep_string=""
 
-if [[ "$USE_QUIJOTE_TAILS" == "true" ]]; then
+# if [[ "$USE_QUIJOTE_TAILS" == "true" ] & [ "$USE_SOBOL" == "false"]]; then
+if [[ "$USE_QUIJOTE_TAILS" == "true" && "$USE_SOBOL" == "false" ]]; then
     cumulants_data_log_dir=$(get_log_dir \
         $BASE_LOG_DIR \
         "cumulants_data" 
@@ -142,6 +167,15 @@ if [[ "$USE_QUIJOTE_TAILS" == "true" ]]; then
     export USE_QUIJOTE_TAILS="$USE_QUIJOTE_TAILS"
     export FIDUCIAL_REDUCE="$FIDUCIAL_REDUCE"
     export DEFAULT_RESOLUTION="$DEFAULT_RESOLUTION"
+    export FORCE_FLAT_PRIOR="$FORCE_FLAT_PRIOR"
+    export FORCE_QUIJOTE_PRIOR="$FORCE_QUIJOTE_PRIOR"
+    export NON_GAUSSIAN_TEST="$NON_GAUSSIAN_TEST"
+    export USE_SOBOL="$USE_SOBOL"
+    export PLOT_FISHER_CLIPPED="$PLOT_FISHER_CLIPPED"
+    export FORCE_RECOMPUTE_DATASET="$FORCE_RECOMPUTE_DATASET"
+    export USE_SCALERS="$USE_SCALERS"
+    export BLACKJAX_SAMPLE="$BLACKJAX_SAMPLE"
+    export AFFINE_SAMPLE="$AFFINE_SAMPLE"
 
     echo \"Running cumulants data script\"
 
@@ -156,52 +190,85 @@ else
     echo "Not running cumulants data scripts"
 fi
 
-# Always run PDFs
-pdfs_data_log_dir=$(get_log_dir \
-    $BASE_LOG_DIR \
-    "pdfs_data"
-)
+if [[ "$USE_SOBOL" == "false" ]]; then
+    # Always run PDFs
+    pdfs_data_log_dir=$(get_log_dir \
+        $BASE_LOG_DIR \
+        "pdfs_data"
+    )
 
-pdfs_cmd="\
+    pdfs_cmd="\
+    #!/bin/bash
+    #SBATCH --job-name=pdfs_data
+    #SBATCH --output=$OUT_DIR/pdfs_data.out
+    #SBATCH --error=$OUT_DIR/pdfs_data.err
+    #SBATCH --partition=cluster
+    #SBATCH --time=$JOB_TIME
+    #SBATCH --mem=${N_GB}GB
+    #SBATCH --cpus-per-task=$N_CPU
+    #SBATCH --mail-user=jed.homer@physik.lmu.de
+    #SBATCH --mail-type=$MAIL_TYPE
+
+    cd /project/ls-gruen/users/jed.homer/sbiaxpdf/cumulants/data/
+    source /project/ls-gruen/users/jed.homer/sbiaxpdf/.venv/bin/activate
+
+    mkdir -p "$pdfs_data_log_dir"
+
+    export N_DATAVECTOR_SEEDS=$N_SEEDS 
+    export N_REPEATED_SBI_SEEDS=$N_SEEDS_GLOBAL 
+    export LOG_DIR="${pdfs_data_log_dir}/"
+    export LOG_LEVEL="$LOG_LEVEL"
+    export RESULTS_DIR="$RESULTS_DIR"
+    export DEFAULT_NDE_TYPE="$DEFAULT_NDE_TYPE"
+    export DEFAULT_N_NDES="$DEFAULT_N_NDES"
+    export FORCE_NOISELESS_DATAVECTOR="$FORCE_NOISELESS_DATAVECTOR"
+    export USE_QUIJOTE_TAILS="$USE_QUIJOTE_TAILS"
+    export FIDUCIAL_REDUCE="$FIDUCIAL_REDUCE"
+    export DEFAULT_RESOLUTION="$DEFAULT_RESOLUTION"
+    export FORCE_FLAT_PRIOR="$FORCE_FLAT_PRIOR"
+    export FORCE_QUIJOTE_PRIOR="$FORCE_QUIJOTE_PRIOR"
+    export NON_GAUSSIAN_TEST="$NON_GAUSSIAN_TEST"
+    export USE_SOBOL="$USE_SOBOL"
+    export PLOT_FISHER_CLIPPED="$PLOT_FISHER_CLIPPED"
+    export FORCE_RECOMPUTE_DATASET="$FORCE_RECOMPUTE_DATASET"
+    export DATA_PROCESS_TYPE_NN="$DATA_PROCESS_TYPE_NN"
+    export USE_PRECISION_NN="$USE_PRECISION_NN"
+    export USE_SCALERS="$USE_SCALERS"
+    export BLACKJAX_SAMPLE="$BLACKJAX_SAMPLE"
+    export AFFINE_SAMPLE="$AFFINE_SAMPLE"
+
+    echo \"Running PDFs data script\"
+
+    python get_pdf_data.py
+    "
+
+    echo "Running PDF data scripts"
+
+    data_job_id=$(echo "$pdfs_cmd" | sbatch | awk '{print $4}')
+    data_deps+=("$data_job_id")
+    data_dep_string=$(IFS=':'; echo "${data_deps[*]}")
+fi
+
+# If not running data jobs, quickly submit an empty job
+# to keep the data_dep_string non-empty
+if [ ${#data_deps[@]} -eq 0 ]; then
+    dummy_job_id=$(sbatch --parsable <<EOF
 #!/bin/bash
-#SBATCH --job-name=pdfs_data
-#SBATCH --output=$OUT_DIR/pdfs_data.out
-#SBATCH --error=$OUT_DIR/pdfs_data.err
-#SBATCH --partition=cluster
-#SBATCH --time=$JOB_TIME
-#SBATCH --mem=${N_GB}GB
-#SBATCH --cpus-per-task=$N_CPU
-#SBATCH --mail-user=jed.homer@physik.lmu.de
-#SBATCH --mail-type=$MAIL_TYPE
+#SBATCH --job-name=dummy
+#SBATCH --time=00:01:00
+#SBATCH --mem=1M
+true
+EOF
+    )
 
-cd /project/ls-gruen/users/jed.homer/sbiaxpdf/cumulants/data/
-source /project/ls-gruen/users/jed.homer/sbiaxpdf/.venv/bin/activate
+    # Initialize dependencies with this dummy job ID
+    data_deps=("$dummy_job_id")
 
-mkdir -p "$pdfs_data_log_dir"
+    # Always build data_dep_string from the array
+    data_dep_string=$(IFS=':'; echo "${data_deps[*]}")
+fi
 
-export N_DATAVECTOR_SEEDS=$N_SEEDS 
-export N_REPEATED_SBI_SEEDS=$N_SEEDS_GLOBAL 
-export LOG_DIR="${pdfs_data_log_dir}/"
-export LOG_LEVEL="$LOG_LEVEL"
-export RESULTS_DIR="$RESULTS_DIR"
-export DEFAULT_NDE_TYPE="$DEFAULT_NDE_TYPE"
-export DEFAULT_N_NDES="$DEFAULT_N_NDES"
-export FORCE_NOISELESS_DATAVECTOR="$FORCE_NOISELESS_DATAVECTOR"
-export USE_QUIJOTE_TAILS="$USE_QUIJOTE_TAILS"
-export FIDUCIAL_REDUCE="$FIDUCIAL_REDUCE"
-export DEFAULT_RESOLUTION="$DEFAULT_RESOLUTION"
-
-echo \"Running PDFs data script\"
-
-python get_pdf_data.py
-"
-
-echo "Running PDF data scripts"
-
-data_job_id=$(echo "$pdfs_cmd" | sbatch | awk '{print $4}')
-data_deps+=("$data_job_id")
-data_dep_string=$(IFS=':'; echo "${data_deps[*]}")
-
+# Run SBI jobs
 for global_seed in $(seq 0 $N_SEEDS_GLOBAL); do
 for FREEZE_FLAG in "--freeze-parameters" "--no-freeze-parameters"; do
     for LINEARISED_FLAG in "--linearised" "--no-linearised"; do
@@ -212,8 +279,15 @@ for FREEZE_FLAG in "--freeze-parameters" "--no-freeze-parameters"; do
                 continue
             fi
 
+            # Skip linearised if not requested
+            if [[ "$RUN_LINEARISED" == false && "$LINEARISED_FLAG" == "--linearised" ]]; then
+                echo "Skipping SBI (linearised)"
+                continue
+            fi
+
             # Skip non-linearised if not requested
             if [[ "$RUN_NONLINEAR" == false && "$LINEARISED_FLAG" == "--no-linearised" ]]; then
+                echo "Skipping SBI (non-linearised)"
                 continue
             fi
 
@@ -291,8 +365,8 @@ $FREEZE_FLAG"
                             cat <<END
 #!/bin/bash
 #SBATCH --job-name=sbi_${global_seed}_${bt_flag}_${l_flag}_${f_flag}_z${z}
-#SBATCH --output=$OUT_DIR/${bt_flag}/${l_flag}/${f_flag}/z${z}/sbi_fixed_%j.out
-#SBATCH --error=$OUT_DIR/${bt_flag}/${l_flag}/${f_flag}/z${z}/sbi_fixed_%j.err
+#SBATCH --output=$OUT_DIR/sbi/${bt_flag}/${l_flag}/${f_flag}/z${z}/sbi_fixed_%j.out
+#SBATCH --error=$OUT_DIR/sbi/${bt_flag}/${l_flag}/${f_flag}/z${z}/sbi_fixed_%j.err
 #SBATCH --partition=cluster
 #SBATCH --time=$JOB_TIME
 #SBATCH --mem=${N_GB}GB
@@ -315,6 +389,17 @@ export FORCE_NOISELESS_DATAVECTOR="$FORCE_NOISELESS_DATAVECTOR"
 export USE_QUIJOTE_TAILS="$USE_QUIJOTE_TAILS"
 export FIDUCIAL_REDUCE="$FIDUCIAL_REDUCE"
 export DEFAULT_RESOLUTION="$DEFAULT_RESOLUTION"
+export FORCE_FLAT_PRIOR="$FORCE_FLAT_PRIOR"
+export FORCE_QUIJOTE_PRIOR="$FORCE_QUIJOTE_PRIOR"
+export NON_GAUSSIAN_TEST="$NON_GAUSSIAN_TEST"
+export USE_SOBOL="$USE_SOBOL"
+export PLOT_FISHER_CLIPPED="$PLOT_FISHER_CLIPPED"
+export FORCE_RECOMPUTE_DATASET="$FORCE_RECOMPUTE_DATASET"
+export DATA_PROCESS_TYPE_NN="$DATA_PROCESS_TYPE_NN"
+export USE_PRECISION_NN="$USE_PRECISION_NN"
+export USE_SCALERS="$USE_SCALERS"
+export BLACKJAX_SAMPLE="$BLACKJAX_SAMPLE"
+export AFFINE_SAMPLE="$AFFINE_SAMPLE"
 
 echo "Running sbi script with seed $global_seed and redshift $z"
 $sbi_cmd
@@ -375,8 +460,8 @@ $FREEZE_FLAG"
                             cat <<END
 #!/bin/bash
 #SBATCH --job-name=m_z_${global_seed}_${bt_flag}_${l_flag}_${f_flag}
-#SBATCH --output=$OUT_DIR/${global_seed}/${bt_flag}/${l_flag}/${f_flag}/m_z_%a_%j.out
-#SBATCH --error=$OUT_DIR/${global_seed}/${bt_flag}/${l_flag}/${f_flag}/m_z_%a_%j.err
+#SBATCH --output=$OUT_DIR/multi_z/${global_seed}/${bt_flag}/${l_flag}/${f_flag}/m_z_%a_%j.out
+#SBATCH --error=$OUT_DIR/multi_z/${global_seed}/${bt_flag}/${l_flag}/${f_flag}/m_z_%a_%j.err
 #SBATCH --array=$JOB_ARRAY_STR
 #SBATCH --partition=cluster
 #SBATCH --time=$JOB_TIME
@@ -400,6 +485,17 @@ export FORCE_NOISELESS_DATAVECTOR="$FORCE_NOISELESS_DATAVECTOR"
 export USE_QUIJOTE_TAILS="$USE_QUIJOTE_TAILS"
 export FIDUCIAL_REDUCE="$FIDUCIAL_REDUCE"
 export DEFAULT_RESOLUTION="$DEFAULT_RESOLUTION"
+export FORCE_FLAT_PRIOR="$FORCE_FLAT_PRIOR"
+export FORCE_QUIJOTE_PRIOR="$FORCE_QUIJOTE_PRIOR"
+export NON_GAUSSIAN_TEST="$NON_GAUSSIAN_TEST"
+export USE_SOBOL="$USE_SOBOL"
+export PLOT_FISHER_CLIPPED="$PLOT_FISHER_CLIPPED"
+export FORCE_RECOMPUTE_DATASET="$FORCE_RECOMPUTE_DATASET"
+export DATA_PROCESS_TYPE_NN="$DATA_PROCESS_TYPE_NN"
+export USE_PRECISION_NN="$USE_PRECISION_NN"
+export USE_SCALERS="$USE_SCALERS"
+export BLACKJAX_SAMPLE="$BLACKJAX_SAMPLE"
+export AFFINE_SAMPLE="$AFFINE_SAMPLE"
 
 echo "Running final multi-z script"
 $multi_z_cmd
@@ -452,9 +548,9 @@ $FREEZE_FLAG"
                     figure_job=$(
                         cat <<END
 #!/bin/bash
-#SBATCH --job-name=figure_one
-#SBATCH --output=$OUT_DIR/${global_seed}/${l_flag}/${f_flag}/figure_one_%a_%j.out
-#SBATCH --error=$OUT_DIR/${global_seed}/${l_flag}/${f_flag}/figure_one_%a_%j.err
+#SBATCH --job-name=figure_1_${global_seed}
+#SBATCH --output=$OUT_DIR/figure_1/${global_seed}/${l_flag}/${f_flag}/figure_one_%a_%j.out
+#SBATCH --error=$OUT_DIR/figure_1/${global_seed}/${l_flag}/${f_flag}/figure_one_%a_%j.err
 #SBATCH --array=$JOB_ARRAY_STR
 #SBATCH --partition=cluster
 #SBATCH --time=$JOB_TIME
@@ -478,6 +574,17 @@ export FORCE_NOISELESS_DATAVECTOR="$FORCE_NOISELESS_DATAVECTOR"
 export USE_QUIJOTE_TAILS="$USE_QUIJOTE_TAILS"
 export FIDUCIAL_REDUCE="$FIDUCIAL_REDUCE"
 export DEFAULT_RESOLUTION="$DEFAULT_RESOLUTION"
+export FORCE_FLAT_PRIOR="$FORCE_FLAT_PRIOR"
+export FORCE_QUIJOTE_PRIOR="$FORCE_QUIJOTE_PRIOR"
+export NON_GAUSSIAN_TEST="$NON_GAUSSIAN_TEST"
+export USE_SOBOL="$USE_SOBOL"
+export PLOT_FISHER_CLIPPED="$PLOT_FISHER_CLIPPED"
+export FORCE_RECOMPUTE_DATASET="$FORCE_RECOMPUTE_DATASET"
+export DATA_PROCESS_TYPE_NN="$DATA_PROCESS_TYPE_NN"
+export USE_PRECISION_NN="$USE_PRECISION_NN"
+export USE_SCALERS="$USE_SCALERS"
+export BLACKJAX_SAMPLE="$BLACKJAX_SAMPLE"
+export AFFINE_SAMPLE="$AFFINE_SAMPLE"
 
 echo "Running final figure one script"
 $figure_cmd
@@ -514,8 +621,15 @@ for order_idx_args in "${order_idxs[@]}"; do
         continue
     fi
 
+    # Skip linearised if not requested
+    if [[ "$RUN_LINEARISED" == false && "$LINEARISED_FLAG" == "--linearised" ]]; then
+        echo "Skipping Figure 2 (linearised)"
+        continue
+    fi
+
     # Skip non-linearised if not requested
     if [[ "$RUN_NONLINEAR" == false && "$LINEARISED_FLAG" == "--no-linearised" ]]; then
+        echo "Skipping Figure 2 (non-linearised)"
         continue
     fi
 
@@ -557,9 +671,9 @@ for order_idx_args in "${order_idxs[@]}"; do
 
     sbatch <<END
 #!/bin/bash
-#SBATCH --job-name=figure_two
-#SBATCH --output=$OUT_DIR/${l_flag}/${f_flag}/figure_two_%j.out
-#SBATCH --error=$OUT_DIR/${l_flag}/${f_flag}/figure_two_%j.err
+#SBATCH --job-name=figure_2
+#SBATCH --output=$OUT_DIR/figure_2/${l_flag}/${f_flag}/figure_two_%j.out
+#SBATCH --error=$OUT_DIR/figure_2/${l_flag}/${f_flag}/figure_two_%j.err
 #SBATCH --partition=cluster
 #SBATCH --time=$JOB_TIME
 #SBATCH --mem=${N_GB}GB
@@ -584,6 +698,16 @@ export FORCE_NOISELESS_DATAVECTOR="$FORCE_NOISELESS_DATAVECTOR"
 export USE_QUIJOTE_TAILS="$USE_QUIJOTE_TAILS"
 export FIDUCIAL_REDUCE="$FIDUCIAL_REDUCE"
 export DEFAULT_RESOLUTION="$DEFAULT_RESOLUTION"
+export FORCE_FLAT_PRIOR="$FORCE_FLAT_PRIOR"
+export FORCE_QUIJOTE_PRIOR="$FORCE_QUIJOTE_PRIOR"
+export NON_GAUSSIAN_TEST="$NON_GAUSSIAN_TEST"
+export USE_SOBOL="$USE_SOBOL"
+export PLOT_FISHER_CLIPPED="$PLOT_FISHER_CLIPPED"
+export DATA_PROCESS_TYPE_NN="$DATA_PROCESS_TYPE_NN"
+export USE_PRECISION_NN="$USE_PRECISION_NN"
+export USE_SCALERS="$USE_SCALERS"
+export BLACKJAX_SAMPLE="$BLACKJAX_SAMPLE"
+export AFFINE_SAMPLE="$AFFINE_SAMPLE"
 
 echo "Running final figure two script"
